+++ CHANGELOG.md
@@ -17,6 +17,11 @@
 
 # 6.2.4
 
+### Changes
+
+- New behavior for Search: the `showSelectedOnly` button now sets the search window in background
+- Extends wmsGroupContentTreeItem to take capabilities set in the property bag of the layer into account
+
 ### Fixes
 
 - Fixes a bug in ContentTreeItem where the `removeAction` function would not trigger reactivity, resulting in removed actions being visible